# backend/openapi/dist.yaml
openapi: 3.0.3
info:
  title: Modular JWT Authenticated API
  version: 1.0.0
  description: An example API using JWT imported from a separate file.
servers:
  - url: 'http://localhost:8080/api/v1'
security:
  - BearerAuth: []
paths:
  /auth/login:
    post:
      summary: Authenticate to receive a JWT
      operationId: login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_LoginCredentials'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/register:
    post:
      summary: Register a new user
      operationId: register
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      summary: Logout and invalidate JWT
      operationId: logout
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/check:
    get:
      summary: Check current JWT and return basic info
      operationId: checkAuth
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh-token:
    post:
      summary: Refresh a JWT
      operationId: refresh-token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: def50200...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /templates:
    post:
      summary: Create a new template
      operationId: createTemplate
      tags:
        - Template
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_CreateTemple'
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /templates/list:
    get:
      summary: List all templates for authenticated user
      operationId: listTemplates
      tags:
        - Template
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Templates retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: '1234567890'
                        name:
                          type: string
                          example: Meeting Notes Template
                        icon:
                          type: string
                          example: Calendar
                        content:
                          type: string
                          example: '<h2>Meeting Notes</h2><p><strong>Date:</strong></p>'
                        tags:
                          type: array
                          items:
                            type: string
                          example:
                            - Meeting
                            - Work
                        color:
                          type: string
                          example: '#5a8a7d'
                        user_id:
                          type: string
                          example: user-uuid-123
                        created_at:
                          type: string
                          format: date-time
                          example: '2021-01-01T00:00:00Z'
                        updated_at:
                          type: string
                          format: date-time
                          example: '2021-01-01T00:00:00Z'
                      required:
                        - id
                        - name
                        - icon
                        - content
                        - tags
                        - user_id
                        - created_at
                        - updated_at
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/templates/{id}':
    get:
      summary: Get template by ID
      operationId: getTemplate
      tags:
        - Template
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Template ID
      responses:
        '200':
          description: Template retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/templates/{id}/update':
    put:
      summary: Update template by ID
      operationId: updateTemplate
      tags:
        - Template
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Template ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_UpdateTemplate'
      responses:
        '200':
          description: Template updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/templates/{id}/delete':
    delete:
      summary: Delete template by ID
      operationId: deleteTemplate
      tags:
        - Template
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Template ID
      responses:
        '204':
          description: Template deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter your JWT token in the format: Bearer <token>.
        The token can be obtained from the `/auth/login` endpoint.
    EditToken:
      type: apiKey
      in: header
      name: X-Edit-Token
      description: Public edit token for note collaboration
  schemas:
    Res_AuthTokens:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJI...
        refresh_token:
          type: string
          example: def50200...
    Req_LoginCredentials:
      type: object
      properties:
        username:
          type: string
          example: johndoe
        password:
          type: string
          format: password
          example: strongpassword123
      required:
        - username
        - password
    Req_UserRegistration:
      type: object
      properties:
        username:
          type: string
          example: janedoe
        password:
          type: string
          format: password
          example: anotherstrongpassword
        email:
          type: string
          format: email
          example: jane.doe@example.com
      required:
        - username
        - password
        - email
    User:
      type: object
      properties:
        id:
          type: string
          description: User UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: john.doe
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        avatar:
          type: string
          format: uri
          example: 'https://example.com/avatar.png'
        avatar_url:
          type: string
          format: uri
          example: 'https://example.com/avatar.png'
        status:
          type: string
          enum:
            - active
            - inactive
            - suspended
          example: active
        email_verified:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
      required:
        - id
        - username
        - name
        - email
        - avatar
        - avatar_url
        - status
        - email_verified
...
    Req_CreateTemple:
      type: object
      properties:
        name:
          type: string
          example: Meeting Notes Template
        icon:
          type: string
          example: Calendar
        content:
          type: string
          example: '<h2>Meeting Notes</h2><p><strong>Date:</strong></p>'
        tags:
          type: array
          items:
            type: string
          example:
            - Meeting
            - Work
        color:
          type: string
          example: '#5a8a7d'
      required:
        - name
        - icon
        - content
    Req_UpdateTemplate:
      type: object
      properties:
        name:
          type: string
          example: Meeting Notes Template
        icon:
          type: string
          example: Calendar
        content:
          type: string
          example: '<h2>Meeting Notes</h2><p><strong>Date:</strong></p>'
        tags:
          type: array
          items:
            type: string
          example:
            - Meeting
            - Work
        color:
          type: string
          example: '#5a8a7d'
    Res_DetailTemplate:
      type: object
      properties:
        id:
          type: string
          example: '1234567890'
        name:
          type: string
          example: Meeting Notes Template
        icon:
          type: string
          example: Calendar
        content:
          type: string
          example: '<h2>Meeting Notes</h2><p><strong>Date:</strong></p>'
        tags:
          type: array
          items:
            type: string
          example:
            - Meeting
            - Work
        color:
          type: string
          example: '#5a8a7d'
        user_id:
          type: string
          example: user-uuid-123
        created_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
      required:
        - id
        - name
        - icon
        - content
        - tags
        - user_id
        - created_at
        - updated_at
...
  responses:
    BadRequest:
      description: Bad Request. The request is invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request.
    Unauthorized:
      description: 'Unauthorized. The provided JWT token is missing, invalid, or expired.'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Authentication token is required.
    NotFound:
      description: Not Found. The requested resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found.
    InternalServerError:
      description: Internal Server Error. The server encountered an unexpected condition that prevented it from fulfilling the request.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Internal server error.
...